Added edit simulation, improved new simulation, added important modals

+++ nsi/src/app/components/behavior-model-list/behavior-model-list.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit, ViewChild } from '@angular/core';
 
 import { BehaviorModel } from '../../models/behavior-model';
 import { BehaviorModelService } from '../../services/behavior-model.service';
+import { Simulation } from '../../models/simulation';
+import { SimulationService } from '../../services/simulation.service';
 
 import { MatTableDataSource } from '@angular/material/table';
 import { MatSort, MatSortable } from '@angular/material/sort';
@@ -11,6 +13,7 @@ import { MatDialog, MatDialogRef } from '@angular/material/dialog';
 import { BehaviorModelCopiedModalComponent } from '../behavior-model-copied-modal/behavior-model-copied-modal.component';
 import { BehaviorModelDeleteConfirmModalComponent } from '../behavior-model-delete-confirm-modal/behavior-model-delete-confirm-modal.component';
 import { BehaviorModelDeletedModalComponent } from '../behavior-model-deleted-modal/behavior-model-deleted-modal.component';
+import { BehaviorModelInUseModalComponent} from '../behavior-model-in-use-modal/behavior-model-in-use-modal.component';
 
 @Component({
   selector: 'app-behavior-model-list',
@@ -25,15 +28,18 @@ export class BehaviorModelListComponent implements OnInit {
   public columns = ['name', 'creationDate', 'valid', 'options'];
   public filterInput: string = "";
   private firstSort = true;
+  private simulationBehaviorModels: any; // ARREGLAR TIPO DE DATOS
 
   @ViewChild(MatSort, {static: true}) private sort!: MatSort;
 
   constructor(private _behaviorModelService: BehaviorModelService,
+              private _simulationService: SimulationService,
               private _matDialog: MatDialog) {
   }
 
-  ngOnInit(): void {
+  async ngOnInit() {
     this.getBehaviorModels();
+    this.simulationBehaviorModels = await this._simulationService.getSimulationBehaviorModels().toPromise();
   }
 
   private getBehaviorModels = () => {
@@ -133,20 +139,47 @@ export class BehaviorModelListComponent implements OnInit {
 
   public openDeleteConfirmModal = (_id: string, name: string) => {
 
-    const dialogRef = this._matDialog.open(BehaviorModelDeleteConfirmModalComponent, { width: '30%' , data: { name: name } } );
+    let modelIsInUse = false;
+    let simulationNamesUsingModels: string[] = [];
 
-    const sub = dialogRef.componentInstance.onSubmit.subscribe(() => {
-      dialogRef.close();
-        this._behaviorModelService.deleteBehaviorModel(_id).subscribe(() => {
-        this.openDeleteNotificationModal(name);
-      }, (error: any) => {
-        console.log(error);
-      })
-    });
+    for (let i = 0; i < this.simulationBehaviorModels.length; i++) {
+      if (this.simulationBehaviorModels[i].simulationBehaviorModelId === _id) {
+        modelIsInUse = true;
+        simulationNamesUsingModels.push(this.simulationBehaviorModels[i].simulationName);
+      }
+    }
 
-    dialogRef.afterClosed().subscribe(() => {
-      sub.unsubscribe();
-    });
+    if (modelIsInUse == true) {
+
+      const dialogRef = this._matDialog.open(BehaviorModelInUseModalComponent, { width: '45%' , data: { name: name, simulationNamesUsingModels: simulationNamesUsingModels } } );
+
+      const sub = dialogRef.componentInstance.onSubmit.subscribe(async () => {
+        dialogRef.close();
+        this.simulationBehaviorModels = await this._simulationService.getSimulationBehaviorModels().toPromise();
+      });
+
+      dialogRef.afterClosed().subscribe(() => {
+        sub.unsubscribe();
+      });
+
+    } else {
+
+      const dialogRef = this._matDialog.open(BehaviorModelDeleteConfirmModalComponent, { width: '30%' , data: { name: name } } );
+
+      const sub = dialogRef.componentInstance.onSubmit.subscribe(() => {
+        dialogRef.close();
+          this._behaviorModelService.deleteBehaviorModel(_id).subscribe(() => {
+          this.openDeleteNotificationModal(name);
+        }, (error: any) => {
+          console.log(error);
+        })
+      });
+
+      dialogRef.afterClosed().subscribe(() => {
+        sub.unsubscribe();
+      });
+
+    }
 
   }
